scraper_service/requirements.txt
# Core dependencies
yfinance>=0.2.33
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=4.9.4

# Web framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.6.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0